Improve code coverage
