+++ src/oehen.arguard.test/ArgumentLessThanGuardTest.cs
@@ -8,21 +8,7 @@ namespace oehen.arguard
     public class ArgumentLessThanGuardTest
     {
         [Theory]
-        [InlineData(-1, 0)]
-        [InlineData(2, 3)]
-        [InlineData(-3, 5)]
-        [InlineData(5, 5)]
-        public void ThrowIfIsLessOrEqualThanCompareValue_ShouldThrowArgumentOutOfRangeException_IfArgumentValueIsLessOrEqualThanTheCompareValue(int argument, int compareValue)
-        {
-            var exception = Assert.Throws<ArgumentOutOfRangeException>(() =>
-            {
-                argument.ThrowIfIsLessOrEqualThan(compareValue, nameof(argument));
-            });
-            exception.Message.Should().Be($"Argument is less or equal than {compareValue}. (Parameter 'argument')" +
-                                          Environment.NewLine + $"Actual value was {argument}.");
-        }
-
-        [Theory]
+        [InlineData(200, 3000)]
         [InlineData(-1, 0)]
         [InlineData(2, 3)]
         [InlineData(-3, 5)]
@@ -37,6 +23,7 @@ namespace oehen.arguard
         }
 
         [Theory]
+        [InlineData(1100, 200)]
         [InlineData(11, 2)]
         [InlineData(3, 2)]
         [InlineData(5, 3)]
@@ -46,19 +33,6 @@ namespace oehen.arguard
             var exception = Record.Exception(() =>
             {
                 argument.ThrowIfIsLessThan(compareValue, nameof(argument));
-            });
-            Assert.Null(exception);
-        }
-
-        [Theory]
-        [InlineData(11, 2)]
-        [InlineData(3, 2)]
-        [InlineData(5, 3)]
-        public void ThrowIfIsLessOrEqualThanCompareValue_ShouldNotThrowArgumentOutOfRangeException_IfArgumentValueIsGreaterOrEqualThanTheCompareValue(int argument, int compareValue)
-        {
-            var exception = Record.Exception(() =>
-            {
-                argument.ThrowIfIsLessOrEqualThan(compareValue, nameof(argument));
             });
             Assert.Null(exception);
         }
